print-title: Don't update the <title> element if we can't decide on a title

+++ se/commands/print_title.py
@@ -35,13 +35,14 @@ def print_title() -> int:
 				title = se.formatting.generate_title(xhtml)
 
 				if args.in_place:
-					processed_xhtml = regex.sub(r"<title>(.*?)</title>", f"<title>{title}</title>", xhtml)
-					processed_xhtml = regex.sub(r"<title/>", f"<title>{title}</title>", processed_xhtml)
+					if title != "":
+						processed_xhtml = regex.sub(r"<title>(.*?)</title>", f"<title>{title}</title>", xhtml)
+						processed_xhtml = regex.sub(r"<title/>", f"<title>{title}</title>", processed_xhtml)
 
-					if processed_xhtml != xhtml:
-						file.seek(0)
-						file.write(processed_xhtml)
-						file.truncate()
+						if processed_xhtml != xhtml:
+							file.seek(0)
+							file.write(processed_xhtml)
+							file.truncate()
 				else:
 					print(title)
